Add files via upload
India is the top agricultural country that produces the largest cotton crop producing country in the world. Here, we are showing the largest Kapas production countries in the World with ranking and production (1000 480 lb. Bales). With 28500 (1000 480 lb. Bales) India is the largest cotton producing country . India covers 125.84 lakh hectare area for cotton production. But due to various disease famers faces a large loss in the production. So to help them know about the disease and to provide information about the best suitable pesticides we made a website where we simply have to upload image of leavees and they will getall the informations. We used flask and keras to train our model.

INSTALLATION: -->pip install Flask -->pip install numpy -->pip install os -->pip install keras